Process relevant fields updating
I added two assignments relevant especially for the SHORT proccesses.
I also update the Tasks file with the progress in updating the
setscheduler function. It seems like all the needed changes in this
function were made. Note that we should still make the actual change of
the process policy between the queues (probably in activate_task and
deactivate_task functions)