Handling the case of SHORT process in activate_task
in the function activate_task I added a condition to handle the case of
a SHORT proccess. if a process is short we should set its "array"
variable to point to the appropriate "prio_array_t" and not the regular
rq->active. This way it'll enqueue it to the right place